fix(golden-harness): render at f64, add strict refactor gate, diff generated source
The golden-audio harness could not support the equivalence claims being made
on top of it. Three measured blind spots, found while designing the
verification gate for a second codegen backend:

1. Renders were stored as f32. runner.rs parsed the driver's full `{:.17e}`
   f64 text and immediately downcast it (`interleaved.push(*v as f32)`),
   putting a ~6e-8 relative floor under every comparison. The differences
   that matter for codegen equivalence — FP contraction, libm divergence —
   are ~1 ulp of f64, far below that floor. A C++ backend compiled with
   gcc's default `-ffp-contract=fast` would very likely have passed the gate
   green. Now f64 end-to-end; captures write `.f64le`.

2. `compare` exited 0 on NEGLIGIBLE. That is the right question for a release
   ("did anything audibly change?") and the wrong one for a behaviour-
   preserving refactor ("did anything change at all?"), because NEGLIGIBLE is
   exactly the band a refactor bug hides in. Added `--strict`, which passes
   only IDENTICAL.

3. `compare` never looked at the generated code at all. Identical audio is
   necessary but not sufficient for identical codegen: the five programs
   never enter most of the recovery ladders, so emitted code can change in
   ways no render executes. `--strict` now also requires every stored
   circuit.rs to match.

Demonstrated on the existing baselines: f80aded vs v013-prep reports 186/186
renders bit-identical and 0 changed, while all 41 circuits' generated code
differs — including the addition of `NODE_NAMES` and `dc_op_by_name`, an
entire public API surface, invisible to the audio gate.

Legacy `.f32le` baselines stay readable so existing captures are not orphaned,
but a byte-equal f32 comparison is annotated as unable to support an f64
bit-exactness claim.

No behaviour change to melange itself; this is harness-only.

+++ tools/golden-harness/src/runner.rs
@@ -148,7 +148,13 @@ pub fn rustc(src: &Path, bin: &Path) -> Result<(), String> {
 pub struct RenderOutput {
     pub channels: usize,
     /// Interleaved samples, frames * channels.
-    pub interleaved: Vec<f32>,
+    ///
+    /// Held at f64. The driver prints full `{:.17e}` f64 text and this used to
+    /// be downcast to f32 on the way in, which put a ~6e-8 relative floor under
+    /// every comparison — far above the ~1-ulp-of-f64 differences that FMA
+    /// contraction and libm divergence produce. A cross-language or refactor
+    /// gate built on that floor cannot see the defects it exists to catch.
+    pub interleaved: Vec<f64>,
     pub frames: usize,
 }
 
@@ -219,7 +225,7 @@ pub fn run_render(bin: &Path, input: &[f64], timeout: Duration) -> Result<Render
 
     let text = String::from_utf8_lossy(&stdout_buf);
     let mut channels = 0usize;
-    let mut interleaved: Vec<f32> = Vec::new();
+    let mut interleaved: Vec<f64> = Vec::new();
     let mut frames = 0usize;
     for line in text.lines() {
         let line = line.trim();
@@ -243,7 +249,7 @@ pub fn run_render(bin: &Path, input: &[f64], timeout: Duration) -> Result<Render
             ));
         }
         for v in &vals {
-            interleaved.push(*v as f32);
+            interleaved.push(*v);
         }
         frames += 1;
     }
